tools/puzzle_importer/requirements.txt
pandas>=2.0.0
python-chess>=1.10.0
